requirements.txt
networkx~=2.5
tqdm~=4.50.2
requests~=2.24.0
numpy~=1.19.2
setuptools~=50.3.1
torch~=1.7.1
dgl~=0.6.1
pytorch-lightning~=1.2.10
gdown~=3.12.2
transformers~=4.5.1
segtok~=1.5.10
flair~=0.8.0
pyvis~=0.1.9
hydra-core
pybiopax~=0.0.6
indra~=1.20.0
scikit-learn~=0.23.2